add int tests to coverage report
