+++ .github/workflows/CI.yml
@@ -31,11 +31,11 @@ jobs:
           go-version: '^1.18'
           cache: true
       - name: Run test
-        run:  go test ./... -coverprofile=coverage.out -v
+        run:  make test_unit
       - uses: actions/upload-artifact@v3
         with:
-          name: coverage
-          path: coverage.out
+          name: coverage_unit
+          path: coverage_unit.out
 
   check-style:
     name: Code style
@@ -71,6 +71,10 @@ jobs:
           cache: true
       - name: Run test
         run:  make test_integration
+      - uses: actions/upload-artifact@v3
+        with:
+          name: coverage_int
+          path: coverage_int.out
       - name: Stop containers
         if: always()
         run: docker compose -f "docker/test_db/docker-compose.yml" down
@@ -115,11 +119,16 @@ jobs:
       - uses: actions/checkout@v3
         with:
           fetch-depth: 0
-      - name: Download line coverage report
+      - name: Download unit tests line coverage report
         uses: actions/download-artifact@v3
         with:
-          name: coverage
-          path: coverage.out
+          name: coverage_unit
+          path: coverage_unit.out
+      - name: Download int tests line coverage report
+        uses: actions/download-artifact@v3
+        with:
+          name: coverage_int
+          path: coverage_int.out
       - name: SonarCloud Scan
         uses: sonarsource/sonarcloud-github-action@master
         env:
