+++ Makefile
@@ -1,14 +1,16 @@
+PATHS = $(shell go list ./... | tail -n +2 | grep -v test_integration | grep -v test_e2e)
+
 test_unit:
-	go test ./... -v
+	go test $(PATHS) -coverpkg=./... -coverprofile=coverage_unit.out -v
 
 test_integration:
 	docker compose -f "docker/test_db/docker-compose.yml" up -d
-	go test -tags=integration ./test_integration -v
+	go test ./test_integration -coverpkg=./... -coverprofile=coverage_int.out -v
 
 test_e2e:
 	docker compose -f "docker/test_db/docker-compose.yml" -f "docker/test_api/docker-compose.yml" up -d
 	./docker/wait_for_api.sh 8000/info
-	go test -tags=e2e ./test_e2e -v
+	go test ./test_e2e -v
 
 example_birds:
 	EXAMPLE=birds docker compose -f "docker/api/docker-compose.yml" up
