+++ sonar-project.properties
@@ -3,7 +3,7 @@ sonar.organization=ditrit
 sonar.projectName=badaas
 sonar.host.url=https://sonarcloud.io
 sonar.sources=.
-sonar.exclusions=**/*_test.go,mocks/***,vendor/***
+sonar.exclusions=**/*_test.go,mocks/***,vendor/***,test_integration/***,test_e2e/***,examples/***
 sonar.tests=.
 sonar.test.inclusions=**/*_test.go
-sonar.go.coverage.reportPaths=./coverage.out/coverage.out
+sonar.go.coverage.reportPaths=coverage_unit.out,coverage_int.out
