+++ test_integration/asserts.go
@@ -1,35 +1,14 @@
-package integration_test
+package integrationtests
 
 import (
-	"log"
 	"reflect"
 	"sort"
 
 	"github.com/ditrit/badaas/persistence/models"
 	"github.com/stretchr/testify/suite"
-	"gorm.io/gorm"
 	"gotest.tools/assert"
 	is "gotest.tools/assert/cmp"
 )
-
-var ListOfTables = []any{
-	models.Session{},
-	models.User{},
-	models.Value{},
-	models.Entity{},
-	models.Attribute{},
-	models.EntityType{},
-}
-
-func SetupDB(db *gorm.DB) {
-	// clean database to ensure independency between tests
-	for _, table := range ListOfTables {
-		err := db.Unscoped().Where("1 = 1").Delete(table).Error
-		if err != nil {
-			log.Fatalln("could not clean database: ", err)
-		}
-	}
-}
 
 func EqualList(ts *suite.Suite, expected, actual any) {
 	v := reflect.ValueOf(expected)
