+++ test_integration/eavservice.go
@@ -1,4 +1,4 @@
-package integration_test
+package integrationtests
 
 import (
 	"fmt"
