+++ test_integration/integration_test.go
@@ -1,4 +1,4 @@
-package integration_test
+package integrationtests
 
 import (
 	"path"
